refactor(cli): reorganize for consistency and better separation of concerns
- Remove redundant cli_ prefix from config files
- Create tmux_helpers.go for tmux-related functions
- Move shared infrastructure (loadConfigForCommand, runSetupHooks) to root.go
- Slim down add.go and stack.go by removing moved functions

File changes:
- cli_config_*.go → config_*.go (consistent naming)
- New: tmux_helpers.go (isInTmux, shouldCreateTmuxWindow, etc.)
- root.go: added loadConfigForCommand, runSetupHooks
- add.go: removed moved functions
- stack.go: removed moved functions
