+++ internal/cli/config_list.go
