+++ internal/cli/config_list_test.go
